Refactor project collection/entity methods/properties
